[REFACTOR]: Add path aliases for root-level folders 🧩
Imports had to walk out of src/ through the @/* alias, producing paths
like "@/../../components/mdx/MdxHeading". Map components, lib, data and
routes explicitly so they read as "@/components/mdx/MdxHeading".
