+++ src/pages/404.tsx
@@ -1,5 +1,5 @@
-import Header from "@/../components/Header";
-import Error404 from "@/../components/errors/Error404";
+import Header from "@/components/Header";
+import Error404 from "@/components/errors/Error404";
 
 export default function Custom404() {
   return (
