+++ src/pages/500.tsx
@@ -1,5 +1,5 @@
-import Header from "@/../components/Header";
-import Error500 from "@/../components/errors/Error500";
+import Header from "@/components/Header";
+import Error500 from "@/components/errors/Error500";
 
 export default function Custom404() {
   return (
