+++ src/pages/projects.tsx
@@ -1,5 +1,5 @@
-import Header from "@/../components/Header";
-import AllProjects from "@/../components/projects/AllProjects";
+import Header from "@/components/Header";
+import AllProjects from "@/components/projects/AllProjects";
 
 export default function Projects() {
   return (
