+++ tsconfig.json
@@ -13,6 +13,10 @@
     "jsx": "preserve",
     "incremental": true,
     "paths": {
+      "@/components/*": ["./components/*"],
+      "@/lib/*": ["./lib/*"],
+      "@/data/*": ["./data/*"],
+      "@/routes": ["./routes"],
       "@/*": ["./src/*"]
     }
   },
